requirements.txt
numpy>=1.19.0
pandas>=1.1.0
scikit-learn>=0.24.0
nltk>=3.6.0
spacy>=3.0.0
textblob>=0.15.3
Pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0